Organized api with listeners
