+++ bridge-to-the-moon/index.js
@@ -1,6 +1,5 @@
 const Web3 = require('web3');
 const fs = require('fs');
-const listeners = require('./src/listeners');
 
 const web3 = new Web3(new Web3.providers.HttpProvider("http://localhost:8545"))
 
@@ -11,5 +10,12 @@ const ScryptVerifier = JSON.parse(fs.readFileSync('../build/contracts/ScryptVeri
 let claimManager = web3.eth.contract(ClaimManager.abi).at(addresses.claimManager);
 let scryptVerifier = web3.eth.contract(ScryptVerifier.abi).at(addresses.scryptVerifier);
 
-console.log("Turning on bridge event listeners")
-listeners.turnOnListeners(claimManager, scryptVerifier);
+async function run() {
+	const listeners = await require('./src/listeners')(web3, [claimManager, scryptVerifier]);	
+
+	console.log("Turning on bridge event listeners")
+
+	listeners.turnOnListeners(claimManager, scryptVerifier);
+}
+
+run();
